requirements.txt
asgiref==3.8.1
Django==5.2.3
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
Markdown==3.8.2
sqlparse==0.5.3
typing_extensions==4.14.0
